+++ pkg/resource/deploy/source_eval.go
@@ -1156,21 +1156,50 @@ func requestFromNodeJS(ctx context.Context) bool {
 }
 
 // transformAliasForNodeJSCompat transforms the alias from the legacy Node.js values to properly specified values.
-func transformAliasForNodeJSCompat(alias resource.Alias) resource.Alias {
-	contract.Assertf(alias.URN == "", "alias.URN must be empty")
-	// The original implementation in the Node.js SDK did not specify aliases correctly:
-	//
-	// - It did not set NoParent when it should have, but instead set Parent to empty.
-	// - It set NoParent to true and left Parent empty when both the alias and resource had no Parent specified.
-	//
-	// To maintain compatibility with such versions of the Node.js SDK, we transform these incorrectly
-	// specified aliases into properly specified ones that work with this implementation of the engine:
-	//
-	// - { Parent: "", NoParent: false } -> { Parent: "", NoParent: true }
-	// - { Parent: "", NoParent: true }  -> { Parent: "", NoParent: false }
-	if alias.Parent == "" {
-		alias.NoParent = !alias.NoParent
+func transformAliasForNodeJSCompat(alias *pulumirpc.Alias) *pulumirpc.Alias {
+	switch a := alias.Alias.(type) {
+	case *pulumirpc.Alias_Spec_:
+		// The original implementation in the Node.js SDK did not specify aliases correctly:
+		//
+		// - It did not set NoParent when it should have, but instead set Parent to empty.
+		// - It set NoParent to true and left Parent empty when both the alias and resource had no Parent specified.
+		//
+		// To maintain compatibility with such versions of the Node.js SDK, we transform these incorrectly
+		// specified aliases into properly specified ones that work with this implementation of the engine:
+		//
+		// - { Parent: "", NoParent: false } -> { Parent: "", NoParent: true }
+		// - { Parent: "", NoParent: true }  -> { Parent: "", NoParent: false }
+		spec := &pulumirpc.Alias_Spec{
+			Name:    a.Spec.Name,
+			Type:    a.Spec.Type,
+			Stack:   a.Spec.Stack,
+			Project: a.Spec.Project,
+		}
+
+		switch p := a.Spec.Parent.(type) {
+		case *pulumirpc.Alias_Spec_ParentUrn:
+			if p.ParentUrn == "" {
+				spec.Parent = &pulumirpc.Alias_Spec_NoParent{NoParent: true}
+			} else {
+				spec.Parent = p
+			}
+		case *pulumirpc.Alias_Spec_NoParent:
+			if p.NoParent {
+				spec.Parent = nil
+			} else {
+				spec.Parent = p
+			}
+		default:
+			spec.Parent = &pulumirpc.Alias_Spec_NoParent{NoParent: true}
+		}
+
+		return &pulumirpc.Alias{
+			Alias: &pulumirpc.Alias_Spec_{
+				Spec: spec,
+			},
+		}
 	}
+
 	return alias
 }
 
@@ -1276,6 +1305,10 @@ func (rm *resmon) RegisterResource(ctx context.Context,
 	transformAliases := !req.GetAliasSpecs() && requestFromNodeJS(ctx)
 
 	for _, aliasObject := range req.GetAliases() {
+		if transformAliases {
+			aliasObject = transformAliasForNodeJSCompat(aliasObject)
+		}
+
 		aliasSpec := aliasObject.GetSpec()
 		var alias resource.Alias
 		if aliasSpec != nil {
@@ -1290,9 +1323,6 @@ func (rm *resmon) RegisterResource(ctx context.Context,
 				Project:  aliasSpec.Project,
 				Parent:   parentURN,
 				NoParent: aliasSpec.GetNoParent(),
-			}
-			if transformAliases {
-				alias = transformAliasForNodeJSCompat(alias)
 			}
 		} else {
 			urn, err := resource.ParseURN(aliasObject.GetUrn())