+++ tests/integration/integration_python_acceptance_test.go
@@ -223,6 +223,7 @@ func optsForConstructPython(
 
 func TestConstructComponentConfigureProviderPython(t *testing.T) {
 	t.Parallel()
+	t.Skip("Temporarily skipping test - pulumi/pulumi#15195")
 
 	const testDir = "construct_component_configure_provider"
 	runComponentSetup(t, testDir)
@@ -242,6 +243,7 @@ func TestConstructComponentConfigureProviderPython(t *testing.T) {
 // Regresses https://github.com/pulumi/pulumi/issues/6471
 func TestAutomaticVenvCreation(t *testing.T) {
 	t.Parallel()
+	t.Skip("Temporarily skipping test - pulumi/pulumi#15195")
 
 	// Do not use integration.ProgramTest to avoid automatic venv
 	// handling by test harness; we actually are testing venv